Split spec.
Most of it was testing KkHakijaService directly
without invoking KkHakijaResource.
